+++ frontend/src/Components/footer/Footer.jsx
@@ -37,7 +37,7 @@ const Footer = () => {
             </div>
             <div>
               <p>Resources</p>
-              <Link to="/forum">Forum</Link>
+              <Link to="/blog">Blog</Link>
               <Link to="/newsletter">Newsletter</Link>
               <Link to="/alcohol"> Alcohol Addiction</Link>
             </div>
