+++ frontend/src/Pages/Blog/Blog.jsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import Download from '../../Components/Download/Download'
+
+const Forum = () => {
+  return (
+    <div>
+        <p className='font-[700] text-[40px] text-center mt-[30px] text-blue mb-[20px] '>Forum</p>
+        <section className='block laptop:flex laptop:flex-rol justify-between mx-auto w-[90%]'>
+            
+        </section>
+        <div className="w-full mb-[50px] mt-[90px] tablet:w-[70%] max-w-[1000px] mx-auto">
+          <Download />
+        </div>
+        <br />
+        <br />
+    </div>
+  )
+}
+
+export default Forum
