+++ frontend/src/Routes/CreateRoute.js
@@ -11,7 +11,7 @@ import Faqs from '../Pages/Faq/Faq'
 import Team from '../Pages/Team/Team';
 
 import TermsServices from '../Pages/TermsServices/TermsServices';
-import Forum from '../Pages/Forum/Forum';
+import Blog from '../Pages/Blog/Blog';
 import Alcohol from '../Pages/Alcohol/Alcohol';
 import JobApplication from '../Pages/JobApplication/JobApplication';
 import NewsLetter from '../Pages/NewsLetter/NewsLetter';
@@ -26,7 +26,7 @@ const CreateRoute = () => {
       <Route path='contact' element={<Contact/>} />
       <Route path='faq' element={<Faqs/>} />
       <Route path='team' element={<Team/>} />
-      <Route path='forum' element={<Forum />} />
+      <Route path='blog' element={<Blog />} />
       <Route path='terms_policy' element={<TermsServices />} />
       <Route path='alcohol' element={<Alcohol />} />
       <Route path='careers/job' element={<JobApplication />} />
